41: Fixed workspace preservation and other timing issues.
All of this was caused by the popup closing and cutting off the code before everything was done. I fixed that by putting refocus prior to cleanup and moving cleanup to the background thread. There were some other little issues as well, but now everything seems to be working correctly in every option state.